+++ src/cocotb/_scheduler.py
@@ -53,7 +53,6 @@ from cocotb.triggers import (
     ReadOnly,
     ReadWrite,
     Trigger,
-    _Join,
 )
 
 # Sadly the Python standard logging module is very slow so it's better not to
@@ -382,8 +381,8 @@ class Scheduler:
         if self._terminate:
             return
 
-        elif _Join(task) in self._trigger2tasks:
-            self._react(_Join(task))
+        elif task.complete in self._trigger2tasks:
+            self._react(task.complete)
 
     def _schedule_task_upon(self, task: Task[Any], trigger: Trigger) -> None:
         """Schedule `task` to be resumed when `trigger` fires."""
@@ -518,13 +517,13 @@ class Scheduler:
     def _trigger_from_started_task(self, result: Task) -> Trigger:
         if _debug:
             self.log.debug(f"Joining to already running task: {result}")
-        return _Join(result)
+        return result.complete
 
     def _trigger_from_unstarted_task(self, result: Task) -> Trigger:
         self._schedule_task(result)
         if _debug:
             self.log.debug(f"Scheduling unstarted task: {result!r}")
-        return _Join(result)
+        return result.complete
 
     def _trigger_from_any(self, result) -> Trigger:
         """Convert a yielded object into a Trigger instance"""